requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0
python-dotenv==1.0.1

celery==5.4.0
redis==5.2.1
requests==2.32.3
gunicorn==23.0.0

pytest==8.3.4
pytest-django==4.10.0
pytest-cov==6.0.0
coverage==7.6.10

flake8==7.1.1
flake8-bugbear==24.12.12
flake8-comprehensions==3.16.0

